move SecurifyError
