+++ src/main/java/ch/securify/SolidityResult.java
@@ -1,8 +1,7 @@
 package ch.securify;
 
 import java.util.TreeMap;
-
-import ch.securify.model.SecurifyError;
+import ch.securify.analysis.SecurifyError;
 
 public class SolidityResult {
     TreeMap<String, SmallPatternResult> results = new TreeMap<>();
